// frontend/src/app/models/Cliente.ts
export class Cliente {
  id?: number;
  nome?: string;
  logradouro?: string;
  numero?: string;
  complemento?: string;
  ddi1?: number;
  ddd1?: number;
  numero1?: number;
  ddi2?: number;
  ddd2?: string;
  numero2?: number;
  email?: string;
  cnpj?: string | null;
  inscricaoEstadual?: string | null;
  contato?: string | null;
  tipoCliente?: string;
  razaoSocial?: string;
}